ecommerce/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'categories', views.CategoryViewSet, basename='category')
router.register(r'products', views.ProductViewSet, basename='product')
router.register(r'contact', views.ContactViewSet)  # Add this line
router.register(r'orders', views.OrderViewSet, basename='order')

urlpatterns = [
    path('', include(router.urls)),
    path('products/by-category/<slug:slug>/', views.ProductsByCategoryView.as_view(), name='products-by-category'),

]